bdmesg can be used in Big Sur and other macOS, by Chris1111

+++ rEFIt_UEFI/Platform/FixBiosDsdt.cpp
@@ -1805,11 +1805,11 @@ UINT32 FixAny (UINT8* dsdt, UINT32 len, const XBuffer<UINT8> ToFind, const XBuff
 
 //    MsgLog(" (%X)", adr);
     found = TRUE;
-    len = move_data(adr + i, dsdt, len, (INT32)sizeoffset); // safe cast, sizeoffset > MAX_INT32
+    len = move_data(adr + i, dsdt, len, (INT32)sizeoffset); // safe cast, sizeoffset < MAX_INT32
     CopyMem(dsdt + adr + i, ToReplace.data(), ToReplace.size());
-    len = CorrectOuterMethod(dsdt, len, adr + i - 2, (INT32)sizeoffset); // safe cast, sizeoffset > MAX_INT32
-    len = CorrectOuters(dsdt, len, adr + i - 3, (INT32)sizeoffset); // safe cast, sizeoffset > MAX_INT32
-    i += adr + ToReplace.size();
+    len = CorrectOuterMethod(dsdt, len, adr + i - 2, (INT32)sizeoffset); // safe cast, sizeoffset < MAX_INT32
+    len = CorrectOuters(dsdt, len, adr + i - 3, (INT32)sizeoffset); // safe cast, sizeoffset < MAX_INT32
+    i += (UINT32)(adr + ToReplace.size()); // if there is no bug before, it should be safe cast.
   }
   MsgLog(" ]\n"); //should not be here
   return len;
@@ -5251,7 +5251,7 @@ void RenameDevices(UINT8* table)
   if ( gSettings.DeviceRenameCount <= 0 ) return; // to avoid message "0 replacement"
 
   INTN i;
-  INTN k=0; // Cland complain about possible use uninitialised. Not true, but I don't like warnings.
+  INTN k=0; // Clang complain about possible use uninitialised. Not true, but I don't like warnings.
   UINTN index;
   INTN size;
   UINTN len = ((EFI_ACPI_DESCRIPTION_HEADER*)table)->Length;
